extract BlockRenderViewMesh superclass from WorldMesh for rendering not-quite-worlds without entities
